move UID to own file
